+++ src/datasources/alarms/constants/DefaultQueries.constants.ts
@@ -1,6 +1,7 @@
 import { AlarmsCountQuery } from '../types/AlarmsCount.types';
-import { ListAlarmsQuery } from '../types/ListAlarms.types';
-import { QueryType } from '../types/types';
+import { AlarmsProperties, ListAlarmsQuery } from '../types/ListAlarms.types';
+import { AlarmsVariableQuery, QueryType } from '../types/types';
+import { DEFAULT_QUERY_EDITOR_DESCENDING, DEFAULT_QUERY_EDITOR_TAKE } from './AlarmsQueryEditor.constants';
 
 export const DEFAULT_QUERY_TYPE: QueryType = QueryType.ListAlarms;
 
@@ -10,4 +11,18 @@ export const defaultAlarmsCountQuery: Omit<AlarmsCountQuery, 'refId'> = {
 
 export const defaultListAlarmsQuery: Omit<ListAlarmsQuery, 'refId'> = {
   filter: '',
+  properties: [
+    AlarmsProperties.displayName,
+    AlarmsProperties.currentSeverityLevel,
+    AlarmsProperties.occurredAt,
+    AlarmsProperties.source,
+    AlarmsProperties.state,
+    AlarmsProperties.workspace,
+  ]
+};
+
+export const defaultListAlarmsVariableQuery: Omit<AlarmsVariableQuery, 'refId'> = {
+  filter: '',
+  take: DEFAULT_QUERY_EDITOR_TAKE,
+  descending: DEFAULT_QUERY_EDITOR_DESCENDING,
 };